Fix bug in Irvin's generator (int -> long long)
Part of #17

+++ partition/generator.cpp
@@ -12,7 +12,7 @@ protected:
     int N, A, B;
     vector<int> D;
 
-    int result;
+    long long result;
 
     void Config() {
         setSlug("partition");
